Refactor token endpoint to add interface for IDP upstream refresh
